+++ src/main/java/org/geowe/client/local/layermanager/ChangeSelectedWMSLayerListener.java
@@ -22,8 +22,14 @@
  */
 package org.geowe.client.local.layermanager;
 
-import org.gwtopenmaps.openlayers.client.layer.Vector;
+import org.gwtopenmaps.openlayers.client.layer.WMS;
 
-public interface ChangeSelectedLayerListener {
-	public void onChange(Vector layer);	
+/**
+ * Representa al responsable de notificar cuando ha cambiado la capa WMS seleccionada en el árbol de capas 
+ *
+ *@autor jose@geowe.org
+ *@since 30/08/2016
+ */
+public interface ChangeSelectedWMSLayerListener {
+	public void onChange(WMS layer);	
 }