+++ src/main/java/org/geowe/client/local/main/tool/map/catalog/LayerLoader.java
@@ -55,6 +55,8 @@ public final class LayerLoader {
 						LayerManagerWidget.VECTOR_TAB, layer);
 			} else {
 				layerManagerWidget.addRaster(layer);
+				layerManagerWidget.setSelectedLayer(
+						LayerManagerWidget.RASTER_TAB, layer);
 			}
 		}
 	}