+++ CHANGELOG.md
@@ -7,13 +7,18 @@ You can see the previous changelog [here] (http://www.geowe.org/index.php?id=cha
 ### Added
 - More spatial tools
 
-## [1.1.0]Beta - 2016-08-26
+## [1.1.0]Beta - 2016-08-31
 
 ### Added
-- Se añade al Catálogo la capa WMTS PNOA
+- Added WMTS PNOA layer into Catalog
+- Added WMTS and TMS support
+- Added zoom level info panel
 
 ### Fixed
 - issue [#151](https://github.com/geowe/geowe-core/issues/151)
+- issue [#152](https://github.com/geowe/geowe-core/issues/152)
+- issue [#153](https://github.com/geowe/geowe-core/issues/153)
+- issue [#159](https://github.com/geowe/geowe-core/issues/159)
 
 ## [1.0.5]Beta - 2016-08-19
 version deployed for testing [map.geowe.org](http://map.geowe.org)